chores: move Connection to config package
