+++ internal/config/builder.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/fljdin/dispatch/internal/tasks"
 	"gopkg.in/yaml.v2"
 )
 
@@ -31,7 +30,7 @@ func (cb *ConfigBuilder) WithTask(task YamlTask) *ConfigBuilder {
 	return cb
 }
 
-func (cb *ConfigBuilder) WithDefaultConnection(cnx tasks.Connection) *ConfigBuilder {
+func (cb *ConfigBuilder) WithDefaultConnection(cnx Connection) *ConfigBuilder {
 	cb.config.DefaultConnection = cnx
 	return cb
 }
