+++ internal/config/config_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	. "github.com/fljdin/dispatch/internal/config"
-	. "github.com/fljdin/dispatch/internal/tasks"
 	"github.com/fljdin/dispatch/internal/tasks/actions"
 	"github.com/lithammer/dedent"
 	"github.com/stretchr/testify/require"
